some more cleanups in the udp functions file structure
